fix: if DifferentHash, discover the file again
There is a possibility of TOCTOU bug where the file is modified on the
client side between being discovered by the watcher (when its hash is
computed) and the file actually being sent to the server, resulting in a
different hash despite the copy happening normally.

To avoid running into such issues, the client receiving a
`DifferentHash` will temporarily forget about the offending file and
discover it again later through its watcher as if it were new.