server db: new `AwaitFromClient` process status
This marks a file that the server has heard about from a client once but
has not been copied yet. This enables better logging instead of
reporting a file-not-found error on the first request. This also allows
preparation on the server side before the file is actually sent, which
will be leveraged to implement subfolders in the incoming file bucket.