ssh_tunnel: `setup_tunnel` returns the `TcpStream`
Prior to this commit, the `setup_tunnel` function would return the local
address before the SSH tunnel is actually in place, expecting the caller
to connect a `TcpStream` to the returned address so that the spawned task
in charge of setting up the tunnel would be able to proceed.

This made the control flow unnecessarily hard to follow. This commit has
`setup_tunnel` directly creating the `TcpStream`, and returning it when
the SSH tunnel is in place.